Avoid returning an score of 20 under some situations.
Note that due to the dimensions of the board, and the mechanics of
the game, there can only be one row/column that reports a non-zero
score.

Note, also that if we just add up the column scores, we could end up
with a situation that looks like:

+---+---+---+
| x | x | x |
+---+---+---+
| 0 | x | 0 |
+---+---+---+
| 0 | 0 | x |
+---+---+---+

And there's no reason why that's more valuable that any other winning
position. Thus, avoid returning 20 in that situation, by returning
early if theres' .